Restructure Android cloudsync package
