+++ apps/android/app/src/main/java/com/flashcardsopensourceapp/app/di/AppGraph.kt
@@ -46,19 +46,19 @@ import com.flashcardsopensourceapp.data.local.review.SharedPreferencesReviewPref
 import com.flashcardsopensourceapp.data.local.repository.AiChatRepository
 import com.flashcardsopensourceapp.data.local.repository.AutoSyncEventRepository
 import com.flashcardsopensourceapp.data.local.repository.CardsRepository
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.CloudIdentityResetCoordinator
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.CloudGuestSessionCoordinator
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.CloudOperationCoordinator
+import com.flashcardsopensourceapp.data.local.repository.cloudsync.account.CloudIdentityResetCoordinator
+import com.flashcardsopensourceapp.data.local.repository.cloudsync.guest.CloudGuestSessionCoordinator
+import com.flashcardsopensourceapp.data.local.repository.cloudsync.runtime.CloudOperationCoordinator
 import com.flashcardsopensourceapp.data.local.repository.CloudAccountRepository
 import com.flashcardsopensourceapp.data.local.repository.DecksRepository
 import com.flashcardsopensourceapp.data.local.repository.ai.LocalAiChatRepository
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.LocalCloudAccountRepository
+import com.flashcardsopensourceapp.data.local.repository.cloudsync.account.LocalCloudAccountRepository
 import com.flashcardsopensourceapp.data.local.repository.LocalCardsRepository
 import com.flashcardsopensourceapp.data.local.repository.LocalDecksRepository
 import com.flashcardsopensourceapp.data.local.repository.progress.LocalProgressCacheStore
 import com.flashcardsopensourceapp.data.local.repository.progress.LocalProgressRepository
 import com.flashcardsopensourceapp.data.local.repository.LocalReviewRepository
-import com.flashcardsopensourceapp.data.local.repository.cloudsync.LocalSyncRepository
+import com.flashcardsopensourceapp.data.local.repository.cloudsync.sync.LocalSyncRepository
 import com.flashcardsopensourceapp.data.local.repository.LocalWorkspaceRepository
 import com.flashcardsopensourceapp.data.local.repository.ProgressRepository
 import com.flashcardsopensourceapp.data.local.repository.ReviewRepository
