+++ apps/android/data/local/src/main/java/com/flashcardsopensourceapp/data/local/repository/cloudsync/sync/LocalOutboxTransactions.kt
@@ -1,4 +1,4 @@
-package com.flashcardsopensourceapp.data.local.repository.cloudsync
+package com.flashcardsopensourceapp.data.local.repository.cloudsync.sync
 
 import androidx.room.withTransaction
 import com.flashcardsopensourceapp.data.local.cloud.CloudPreferencesStore
